Code Analyzer
Implemented the code analyzer tool using C# which could be used as a
plugin for eclipse IDE to find the class hierarchy. It can be also used
in a compiler, to find the dependency of various classes, structures
etc. Since the tool evaluates the complexity of functions too, it could
be used as a complexity evaluator.